Made falling deaths faster in main play area
Because the apple area is so far below the starting area, deaths due to
falling in the main area take a long time to reset because the player
has to fall so far before dying. I added a collision box to extend the
death-zone up a bit underneath the main area just to make falling deaths
there a little snappier.