Improved apple platform collision detection a bit
There was a possibility of triggering the platform movement just right
so you get left behind. This should take care of that case.

I also noticed the platform could be primed before landing on it by
jumping over it. I will have to keep an eye on this and try to provide a
solution later.