Fixed coyote jump not being the same for both directions
The jump allowance when moving left was shorter than the allowance when
moving right. Now that the character speed has been decreased, the one
missing pixel on the left jump was breaking the ability for the
character to clear a gap of 4 blocks while the right jump could clear
that same gap.