Removed resetting the apple on player death
Having to make it all the way back to the Top of the World in one chance
is simply too difficult for most players. There is so little room for
error that it quickly becomes frustrating. Now, the game functions in
the same way all the way through, keeping players focused on improving
instead of fighting the game.