Removed whoosh images from player animations
I do not like how the whooshes travel with the player as he moves after
initiating the command. Instead, I want to spawn the whoosh as an object
in the world when the player takes their action then quickly fade them
out. This way, the whoosh stays in a place that makes sense in the
context of the world.