Disabled climbing when player is directly under blocks
The blocks keep the player from climbing at all, but the climb checker
thought there was a climb possible because the ceiling was only one
block thick, not realizing there was a block over the player. I checked
the block over the player to prevent this.

I had to make an exception to the check for platforms since the player
can climb from underneath them regardless of how close to the player's
head they are. Platforms do not prevent climbing, so they should not be
treated as such.